managed-agentcore/requirements.txt
# Python dependencies for AgentCore Runtime Container
# Synced with pyproject.toml

# Core packages
ipykernel==7.1.0
jupyter==1.1.1
boto3==1.40.75

# Strands SDK and AgentCore
strands-agents==1.17.0
strands-agents-tools==0.2.16
bedrock-agentcore>=1.0.6
bedrock-agentcore-starter-toolkit>=0.1.33

# MCP and utilities
mcp==1.21.0
streamlit==1.51.0
requests-toolbelt==1.0.0
sagemaker==2.254.1
awscli==1.42.75
termcolor==3.2.0

# Observability
aws-opentelemetry-distro==0.12.0

# Data visualization
matplotlib==3.10.7
matplotlib-inline==0.2.1
seaborn==0.13.2
lovelyplots>=1.0.2
plotly==6.4.0
koreanize-matplotlib>=0.1.1

# Data processing
pandas==2.3.3
numpy>=1.26.4
scikit-learn==1.7.2

# Document generation
python-docx==1.2.0
weasyprint==66.0
reportlab==4.4.4
markdown==3.10
fpdf2==2.8.5

# Other utilities
python-dotenv==1.2.1
pillow>=11.2.1
langchain==1.0.7
